correction erreur cast trajectoire

+++ trajectoire.c
@@ -1,5 +1,6 @@
 #include "odometrie.h"
 #include <math.h> //utiliser un tableau pour acos ??
+#include <stdio.h> //TODO : à virer
 
 //ici le calcul de la trajectoire à effectuer
 
@@ -44,11 +45,9 @@ void consigne_new_xy_relatif(int x_voulu, int y_voulu)
 	//x et y sont relatifs mais l'orientation ne change pas avec celle du robot
 
 	int new_delta=(int)sqrt((double)(x_voulu*x_voulu+y_voulu*y_voulu)); //voir si pas meilleur moyen
-	//pas bon à revoir type de variable mauvais ou alors passer à un angle en degré -> passage en milliradian depuis
 	int sgn_x=(x_voulu > 0) - (x_voulu < 0);
-	//sgn_x=1;
-	int theta_voulu=(int)(1000.0*acos((double)(y_voulu/new_delta))*(-1*sgn_x)); //voir si pas meilleur moyen (tableau ?)
-	printf("%d\n", theta_voulu);
+	int theta_voulu=(int)(1000.0*acos((double)(y_voulu)/(double)(new_delta))*(-1.0*sgn_x)); //voir si pas meilleur moyen (tableau ?)
+	printf("%d\n",theta_voulu);
 	int new_alpha=theta_voulu-get_theta_actuel();
 	//acos retourne et prend un double à priori
 	consigne_new_alpha_delta(new_alpha,new_delta);
